requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
mpmath==1.3.0
numpy==2.2.6
openai==1.104.2
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pytesseract==0.3.13
sniffio==1.3.1
sympy==1.14.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
